Extract large module
